backend/requirements.txt
# Backend Requirements for Development Portal Assistant

# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Environment and Configuration
python-dotenv==1.0.0

# HTTP and Data Handling
requests==2.31.0
pydantic==2.5.0
httpx==0.25.2

# Vector Database and Embeddings
pinecone-client==2.2.4
sentence-transformers==2.2.2
torch==2.1.1
numpy==1.24.3

# Text Processing and Parsing
beautifulsoup4==4.12.2
html2text==2020.1.16
markdownify==0.11.6
lxml==4.9.3

# Image Processing
Pillow==10.1.0
google-generativeai==0.3.2

# AI and ML
openai==1.3.6
transformers==4.36.0

# Data Science and Utilities
pandas==2.1.3
tqdm==4.66.1

# Frontend
streamlit==1.28.1

# Testing and Development
pytest==7.4.3
pytest-asyncio==0.21.1

# Additional utilities
python-magic==0.4.27
aiofiles==23.2.1
asyncio-throttle==1.0.2

# Vector Database
pinecone

# Embeddings and ML
scikit-learn>=1.3.0

# Text Processing
python-docx>=0.8.11
PyPDF2>=3.0.0
pdfplumber>=0.9.0

# Async and Utilities
aiohttp>=3.8.0

# Logging and Monitoring
structlog>=23.1.0

# Optional: Advanced OCR
easyocr>=1.7.0

# Optional: For better logging
colorama>=0.4.6 